requirements.txt
absl-py==0.8.1
astor==0.8.1
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2020.12.5
cffi==1.14.5
Click==7.0
dataclasses==0.8
DateTime==4.3
enum34==1.1.6
Flask==1.1.1
funcsigs==1.0.2
gast==0.4.0
grpcio==1.36.1
gunicorn==19.9.0
h5py==2.10.0
html5lib==0.9999999
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.5
Pillow==6.2.1
protobuf==3.10.0
pycparser==2.20
PySoundFile==0.9.0.post1
pytz==2021.1
PyYAML==5.1.2
scipy==1.2.0
six==1.12.0
SpeechRecognition==3.8.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==2.0.1
zope.interface==5.3.0
